requirements-indexing.txt
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
requests==2.31.0
